feat: moved processed data to virtual module and added github action for Pages
