+++ webpack.common.js
@@ -1,8 +1,8 @@
-const webpack = require('webpack');
-const path = require('path');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
+import webpack from 'webpack';
+import path from 'path';
+import HtmlWebpackPlugin from 'html-webpack-plugin';
 
-module.exports = {
+export default {
     entry: './src/js/index.js',
     module: {
         rules: [
@@ -13,6 +13,12 @@ module.exports = {
             {
                 test: /\.(png|svg|jpg|jpeg|gif)$/i,
                 type: 'asset/resource',
+            },
+            {
+                // This rule specifically targets the resolved path of the loader script.
+                // It ensures the loader is processed as a JavaScript module.
+                test: path.resolve(process.cwd(), 'scripts/preprocessor-loader.js'),
+                use: [path.resolve(process.cwd(), 'scripts/preprocessor-loader.js')],
             },
         ],
     },
@@ -30,14 +36,20 @@ module.exports = {
                 'og:title': { property: 'og:title', content: 'Ingress Shards' },
             }
         }),
+        new webpack.NormalModuleReplacementPlugin(
+            /processed_shard_series.json$/,
+            (resource) => {
+                resource.request = path.resolve(process.cwd(), 'scripts/preprocessor-loader.js');
+            }
+        ),
     ],
     resolve: {
         alias: {
-            'leaflet.motion': path.resolve(__dirname, 'node_modules/leaflet.motion/dist/leaflet.motion.min.js')
+            'leaflet.motion': path.resolve(process.cwd(), 'node_modules/leaflet.motion/dist/leaflet.motion.min.js'),
         }
     },
     output: {
-        path: path.resolve(__dirname, 'dist'),
+        path: path.resolve(process.cwd(), 'dist'),
         filename: '[name].bundle.js',
         assetModuleFilename: 'images/[hash][ext][query]',
         clean: true,
